backend/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
black==22.6.0
click==8.1.3
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
django-livereload-server==0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
tornado==6.2
